+++ tapo-py/src/api/discovery.rs
@@ -1,5 +1,9 @@
 mod device_discovery;
+mod device_discovery_raw;
+mod discovery_raw_result;
 mod discovery_result;
 
 pub use device_discovery::*;
+pub use device_discovery_raw::*;
+pub use discovery_raw_result::*;
 pub use discovery_result::*;